[ARC-584] Extract project format to separate module
branch: feature/ARC-584-discover-service-dependencies
