+++ src/mpyl/cli/commands/projects/__init__.py
@@ -1,0 +1,1 @@
+"""Project related command implementations"""
